Inital Commit of the new ODeSA_Scan
It compiles (with a few warnings from the PulseAnalysis.h about "sizeof" and pointers, but this exists in the old version as well, so im leaving them for now. (TO BE fixed later once the rest of the scan is rewritten) 

It properly reads the CMD line flags (easy part), and couts them.